+++ artifacts/AdditionalSetupArtifacts.ps1
@@ -120,11 +120,11 @@ Write-Host "=== Additional Setup ==="
 if (Test-Path "$serviceTierFolder") {
     if (Test-Path "$serviceTierFolder\Microsoft.Dynamics.Nav.Management.psm1") {
         Write-Host "Import Management Utils from $serviceTierFolder\Microsoft.Dynamics.Nav.Management.psm1"
-        Import-Module "$serviceTierFolder\Microsoft.Dynamics.Nav.Management.psm1" -Force -ErrorAction SilentlyContinue -DisableNameChecking
+        Import-Module "$serviceTierFolder\Microsoft.Dynamics.Nav.Management.psm1" -Force -ErrorAction SilentlyContinue -DisableNameChecking 2>$null
     }
     else {
         Write-Host "Import Management Utils from $serviceTierFolder\Microsoft.Dynamics.Nav.Management.dll"
-        Import-Module "$serviceTierFolder\Microsoft.Dynamics.Nav.Management.dll" -Force -ErrorAction SilentlyContinue -DisableNameChecking
+        Import-Module "$serviceTierFolder\Microsoft.Dynamics.Nav.Management.dll" -Force -ErrorAction SilentlyContinue -DisableNameChecking 2>$null
     }
     if (Test-Path "$serviceTierFolder\Microsoft.Dynamics.Nav.Apps.Management.psd1") {
         Write-Host "Import App Management Utils from $serviceTierFolder\Microsoft.Dynamics.Nav.Apps.Management.psd1"
@@ -137,7 +137,7 @@ if (Test-Path "$serviceTierFolder") {
 }
 if (Test-Path "$roleTailoredClientFolder\Microsoft.Dynamics.Nav.Ide.psm1") {
     Write-Host "Import Nav IDE from $roleTailoredClientFolder\Microsoft.Dynamics.Nav.Ide.psm1"
-    Import-Module "$roleTailoredClientFolder\Microsoft.Dynamics.Nav.Ide.psm1" -Force -ErrorAction SilentlyContinue -DisableNameChecking
+    Import-Module "$roleTailoredClientFolder\Microsoft.Dynamics.Nav.Ide.psm1" -Force -ErrorAction SilentlyContinue -DisableNameChecking 2>$null
 }
 
 if ((Test-Path 'c:\run\cosmo.compiler.helper.psm1') -and ($env:IsBuildContainer)) {