+++ artifacts/OverridesHandling/Expand-Archive.ps1
@@ -1,4 +1,5 @@
 function Expand-Archive() {
+    # Must be a simple function for correct splatting
     try {
         $previousProgressPreference = $global:ProgressPreference
         $global:ProgressPreference = 'SilentlyContinue'