+++ artifacts/OverridesHandling/Invoke-WebRequest.ps1
@@ -1,4 +1,5 @@
 function Invoke-WebRequest() { 
+    # Must be a simple function for correct splatting
     try {
         $previousProgressPreference = $global:ProgressPreference
         $global:ProgressPreference = 'SilentlyContinue'